feat: add qknn_regressor and generalize qknn in a base class
* implement a neighbors regressor in a quantum fashion
* generalize knn in a base class
* rename qknn.py to qknn_classifier.py
