+++ qlkit/algorithms/__init__.py
@@ -1,11 +1,12 @@
 from .quantum_estimator import QuantumEstimator
-from .qknn import QKNeighborsClassifier
+from .qknn import QKNeighborsClassifier, QKNeighborsRegressor
 from .qsvm import QSVClassifier
 from .qkmeans import QKMeans
 
 __all__ = [
     "QuantumEstimator",
     "QKNeighborsClassifier",
+    "QKNeighborsRegressor",
     "QSVClassifier",
     "QKMeans"
 ]
