+++ qlkit/algorithms/qknn/__init__.py
@@ -1,5 +1,7 @@
-from .qknn import QKNeighborsClassifier
+from .qknn_classifier import QKNeighborsClassifier
+from .qknn_regressor import QKNeighborsRegressor
 
 __all__ = [
-    "QKNeighborsClassifier"
+    "QKNeighborsClassifier",
+    "QKNeighborsRegressor"
 ]
